feat: AppEventHandler + remove annotation injection in Root
