+++ p2p_wallet/Scenes/Root/Root.ViewModel.swift
@@ -17,19 +17,13 @@ protocol RootViewModelType {
     var resetSignal: Signal<Void> {get}
     
     func reload()
-    func logout()
     func finishSetup()
-}
-
-protocol CreateOrRestoreWalletHandler {
-    func creatingWalletDidComplete(phrases: [String]?, derivablePath: SolanaSDK.DerivablePath?, name: String?)
-    func restoringWalletDidComplete(phrases: [String]?, derivablePath: SolanaSDK.DerivablePath?, name: String?)
-    func creatingOrRestoringWalletDidCancel()
 }
 
 extension Root {
     class ViewModel {
         // MARK: - Dependencies
+        private var appEventHandler: AppEventHandlerType = Resolver.resolve()
         private let storage: AccountStorageType & PincodeStorageType & NameStorageType = Resolver.resolve()
         private let analyticsManager: AnalyticsManagerType = Resolver.resolve()
         private let notificationsService: NotificationsServiceType = Resolver.resolve()
@@ -38,7 +32,11 @@ extension Root {
         private let disposeBag = DisposeBag()
         private var isRestoration = false
         private var showAuthenticationOnMainOnAppear = true
-        private var resolvedName: String?
+        
+        // MARK: - Initializer
+        init() {
+            bind()
+        }
         
         deinit {
             debugPrint("\(String(describing: self)) deinited")
@@ -50,6 +48,13 @@ extension Root {
         private let resetSubject = PublishRelay<Void>()
         
         // MARK: - Actions
+        private func bind() {
+            appEventHandler.delegate = self
+            appEventHandler.isLoadingDriver
+                .drive(isLoadingSubject)
+                .disposed(by: disposeBag)
+        }
+        
         func reload() {
             // signal VC to prepare for reseting
             resetSubject.accept(())
@@ -101,50 +106,35 @@ extension Root.ViewModel: RootViewModelType {
     }
 }
 
-extension Root.ViewModel: DeviceOwnerAuthenticationHandler {
-    func requiredOwner(onSuccess: (() -> Void)?, onFailure: ((String?) -> Void)?) {
-        let myContext = LAContext()
-        
-        var error: NSError?
-        guard myContext.canEvaluatePolicy(.deviceOwnerAuthentication, error: &error) else {
-            DispatchQueue.main.async {
-                onFailure?(errorToString(error))
-            }
-            return
-        }
-        
-        myContext.evaluatePolicy(.deviceOwnerAuthentication, localizedReason: L10n.confirmItSYou) { (success, error) in
-            guard success else {
-                DispatchQueue.main.async {
-                    onFailure?(errorToString(error))
-                }
-                return
-            }
-            DispatchQueue.main.sync {
-                onSuccess?()
-            }
-        }
+extension Root.ViewModel: AppEventHandlerDelegate {
+    func createWalletDidComplete() {
+        isRestoration = false
+        navigationSubject.accept(.onboarding)
+        analyticsManager.log(event: .setupOpen(fromPage: "create_wallet"))
     }
-}
-
-extension Root.ViewModel: ChangeNetworkResponder {
-    func changeAPIEndpoint(to endpoint: SolanaSDK.APIEndPoint) {
-        Defaults.apiEndPoint = endpoint
-        
+    
+    func restoreWalletDidComplete() {
+        isRestoration = true
+        navigationSubject.accept(.onboarding)
+        analyticsManager.log(event: .setupOpen(fromPage: "recovery"))
+    }
+    
+    func onboardingDidFinish(resolvedName: String?) {
+        let event: AnalyticsEvent = isRestoration ? .setupWelcomeBackOpen: .setupFinishOpen
+        analyticsManager.log(event: event)
+        navigationSubject.accept(.onboardingDone(isRestoration: isRestoration, name: resolvedName))
+    }
+    
+    func userDidChangeAPIEndpoint(to endpoint: SolanaSDK.APIEndPoint) {
         showAuthenticationOnMainOnAppear = false
-        ResolverScope.session.reset()
         reload()
         
         DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) { [weak self] in
             self?.notificationsService.showInAppNotification(.done(L10n.networkChanged))
         }
     }
-}
-
-extension Root.ViewModel: ChangeLanguageResponder {
-    func languageDidChange(to language: LocalizedLanguage) {
-        UIApplication.languageChanged()
-        
+    
+    func userDidChangeLanguage(to language: LocalizedLanguage) {
         showAuthenticationOnMainOnAppear = false
         reload()
         
@@ -153,97 +143,8 @@ extension Root.ViewModel: ChangeLanguageResponder {
             self?.notificationsService.showInAppNotification(.done(languageChangedText))
         }
     }
-}
-
-extension Root.ViewModel: LogoutResponder {
-    func logout() {
-        storage.clearAccount()
-        Defaults.walletName = [:]
-        Defaults.didSetEnableBiometry = false
-        Defaults.didSetEnableNotifications = false
-        Defaults.didBackupOffline = false
-        Defaults.renVMSession = nil
-        Defaults.renVMProcessingTxs = []
-        Defaults.forceCloseNameServiceBanner = false
-        Defaults.shouldShowConfirmAlertOnSend = true
-        Defaults.shouldShowConfirmAlertOnSwap = true
+    
+    func userDidLogout() {
         reload()
     }
-}
-
-extension Root.ViewModel: CreateOrRestoreWalletHandler {
-    func creatingWalletDidComplete(phrases: [String]?, derivablePath: SolanaSDK.DerivablePath?, name: String?) {
-        isRestoration = false
-        resolvedName = name
-        navigationSubject.accept(.onboarding)
-        analyticsManager.log(event: .setupOpen(fromPage: "create_wallet"))
-        saveAccountToStorage(phrases: phrases, derivablePath: derivablePath, name: name)
-    }
-    
-    func restoringWalletDidComplete(phrases: [String]?, derivablePath: SolanaSDK.DerivablePath?, name: String?) {
-        isRestoration = true
-        resolvedName = name
-        navigationSubject.accept(.onboarding)
-        analyticsManager.log(event: .setupOpen(fromPage: "recovery"))
-        saveAccountToStorage(phrases: phrases, derivablePath: derivablePath, name: name)
-    }
-    
-    func creatingOrRestoringWalletDidCancel() {
-        logout()
-    }
-    
-    private func saveAccountToStorage(phrases: [String]?, derivablePath: SolanaSDK.DerivablePath?, name: String?) {
-        guard let phrases = phrases, let derivablePath = derivablePath else {
-            creatingOrRestoringWalletDidCancel()
-            return
-        }
-        
-        isLoadingSubject.accept(true)
-        DispatchQueue.global().async { [weak self] in
-            do {
-                try self?.storage.save(phrases: phrases)
-                try self?.storage.save(derivableType: derivablePath.type)
-                try self?.storage.save(walletIndex: derivablePath.walletIndex)
-                
-                if let name = name {
-                    self?.storage.save(name: name)
-                }
-                
-                DispatchQueue.main.async { [weak self] in
-                    self?.isLoadingSubject.accept(false)
-                }
-            } catch {
-                self?.isLoadingSubject.accept(false)
-                DispatchQueue.main.async { [weak self] in
-                    self?.notificationsService.showInAppNotification(.error(error))
-                    self?.creatingOrRestoringWalletDidCancel()
-                }
-            }
-        }
-    }
-}
-
-extension Root.ViewModel: OnboardingHandler {
-    func onboardingDidCancel() {
-        logout()
-    }
-    
-    @objc func onboardingDidComplete() {
-        let event: AnalyticsEvent = isRestoration ? .setupWelcomeBackOpen: .setupFinishOpen
-        analyticsManager.log(event: event)
-        navigationSubject.accept(.onboardingDone(isRestoration: isRestoration, name: resolvedName))
-    }
-}
-
-private func errorToString(_ error: Error?) -> String? {
-    var error = error?.localizedDescription ?? L10n.unknownError
-    switch error {
-    case "Passcode not set.":
-        error = L10n.PasscodeNotSet.soWeCanTVerifyYouAsTheDeviceSOwner
-    case "Canceled by user.":
-        return nil
-    default:
-        break
-    }
-    return error
 }
